refactor : divide to components, rename
